Refactor Header into component
