+++ src/App.js
@@ -2,32 +2,21 @@ import React, { useState, useEffect } from "react";
 
 import { makeStyles } from "@material-ui/core/styles";
 import {
-  AppBar,
   CircularProgress,
   Container,
   CssBaseline,
   Grid,
-  Tab,
-  Tabs,
-  Toolbar,
   Typography,
-  useMediaQuery,
 } from "@material-ui/core";
-import { BrowserRouter, Link, Redirect, Route, Switch } from "react-router-dom";
+import { BrowserRouter, Redirect, Route, Switch } from "react-router-dom";
 
+import Header from "./Header";
 import Projects from "./Projects";
 
 const useStyles = makeStyles((theme) => ({
   container: {
     paddingTop: theme.spacing(4),
     paddingBottom: theme.spacing(4),
-  },
-  nav: {
-    color: theme.palette.getContrastText(theme.palette.info.dark),
-    backgroundColor: theme.palette.info.dark,
-  },
-  fullHeightTabs: {
-    ...theme.mixins.toolbar,
   },
   appBarSpacer: theme.mixins.toolbar,
   fullHeight: {
@@ -68,8 +57,6 @@ function App() {
     setFavourites(update);
   };
 
-  const centerNav = useMediaQuery("(min-width:780px)");
-
   const classes = useStyles();
 
   return (
@@ -103,42 +90,7 @@ function App() {
           </Container>
         ) : (
           <div>
-            <AppBar className={classes.nav}>
-              <Toolbar>
-                <Grid container alignItems="center" justify="space-between">
-                  <Grid item xs>
-                    <Typography variant="h6">2021 Part IV Projects</Typography>
-                  </Grid>
-                  <Grid item>
-                    <Route
-                      path="/"
-                      render={({ location }) => (
-                        <Tabs
-                          value={location.pathname}
-                          className={classes.fullHeightTabs}
-                        >
-                          <Tab
-                            label="Explore"
-                            className={classes.fullHeightTabs}
-                            component={Link}
-                            to="/explore"
-                            value="/explore"
-                          />
-                          <Tab
-                            label="Rank"
-                            className={classes.fullHeightTabs}
-                            component={Link}
-                            to="/rank"
-                            value="/rank"
-                          />
-                        </Tabs>
-                      )}
-                    />
-                  </Grid>
-                  {centerNav && <Grid item xs />}
-                </Grid>
-              </Toolbar>
-            </AppBar>
+            <Header />
             <main>
               <div className={classes.appBarSpacer} />
               <Container className={classes.container}>
