+++ src/Projects.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 
 import Filter from "./Filter";
 import Project from "./Project";
